Add erp redeem implementation with 60 standard and 4 emergency keys
